+++ cardshifter-core/src/main/java/com/cardshifter/core/game/TCGGame.java
@@ -239,7 +239,7 @@ public class TCGGame extends ServerGame {
         Entity performer = playerFor(client);
 		boolean allowed = action.perform(performer);
 		if (!allowed) {
-			client.sendToClient(new ServerErrorMessage("Action not allowed: " + action));
+			client.sendToClient(new ErrorMessage("Action not allowed: " + action));
 		}
 		
 		sendAvailableActions();
