+++ cardshifter-server/src/main/java/com/cardshifter/server/clients/ClientSocketHandler.java
@@ -16,13 +16,10 @@ import org.apache.log4j.Logger;
 import com.cardshifter.api.ClientIO;
 import com.cardshifter.api.incoming.TransformerMessage;
 import com.cardshifter.api.messages.Message;
-import com.cardshifter.api.outgoing.ServerErrorMessage;
 import com.cardshifter.api.serial.CommunicationTransformer;
 import com.cardshifter.server.model.Server;
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 public class ClientSocketHandler extends ClientIO implements Runnable {
