+++ cardshifter-server/src/main/java/com/cardshifter/server/model/InviteManager.java
@@ -58,7 +58,7 @@ public class InviteManager {
             logger.info("Sending [" + invite + "] to [" + receiver + "]");
             invite.sendInvite(receiver);
         } else {
-            sender.sendToClient(new ServerErrorMessage("You already have a game invitation open."));
+            sender.sendToClient(new ErrorMessage("You already have a game invitation open."));
         }
     }
